Move WelcomeMessage component to home app
