+++ app/packs/src/apps/home/WelcomeMessage.js
